pyproject.toml
[project]
name = "chixia-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "interactions-py>=5.14.0",
    "loguru>=0.7.3",
    "ruff>=0.9.10",
]
